+++ src/dev_config.yaml
@@ -11,7 +11,7 @@ global:
   network_or_hdfs: "network" #whether to load from hdfs or network (local Python)
   # Construction Settings
   load_reference_list: True # the reference list corrections should always be loaded in year 1
-  load_from_feather: True
+  load_from_feather: False
   # Apportionment to sites settings
   apportion_sites: True
 runlog_writer:
