+++ src/mapping/mapping_main.py
@@ -7,6 +7,7 @@ from src.mapping import mapping_helpers as hlp
 from src.mapping.pg_conversion import run_pg_conversion
 from src.mapping.ultfoc_mapping import join_fgn_ownership
 from src.mapping.cellno_mapping import validate_join_cellno_mapper
+from src.mapping.itl_mapping import join_itl_regions
 from src.staging import staging_helpers as stage_hlp
 from src.staging import validation as val
 
@@ -16,6 +17,7 @@ MappingMainLogger = logging.getLogger(__name__)
 def run_mapping(
     full_responses,
     ni_full_responses,
+    postcode_mapper,
     config: dict,
     write_csv: Callable,
     run_id: int,
@@ -71,6 +73,7 @@ def run_mapping(
     full_responses = run_pg_conversion(full_responses, pg_num_alpha, sic_pg_num)
     full_responses = join_fgn_ownership(full_responses, ultfoc_mapper)
     full_responses = validate_join_cellno_mapper(full_responses, cellno_df, config)
+    full_responses = join_itl_regions(full_responses, postcode_mapper)
 
     if ni_full_responses is not None:
         ni_full_responses = hlp.create_additional_ni_cols(ni_full_responses)
@@ -111,4 +114,4 @@ def run_mapping(
     MappingMainLogger.info("Finished Mapping NI QA calculation.")
 
     # return mapped_df
-    return (full_responses, ni_full_responses, itl_mapper, cellno_df)
+    return (full_responses, ni_full_responses, itl_mapper)
