+++ tests/test_utils/test_path_helpers.py
@@ -24,8 +24,12 @@ def config():
             "ni_full_responses_path": "03_northern_ireland/2021/TEST_ni.csv",
             "manual_imp_trim_path": "06_imputation/man_trim/trim_qa.csv",
             "manual_outliers_path": "07_outliers/man_out/man_out.csv",
-            "all_data_construction_file_path": "04_construction/man_con/construction_file.csv",
-            "postcode_construction_file_path": "04_construction/man_con/postcode_construction_file.csv",
+            "all_data_construction_file_path": (
+                "04_construction/man_con/construction_file.csv"
+            ),
+            "postcode_construction_file_path": (
+                "04_construction/man_con/postcode_construction_file.csv"
+            ),
             "construction_file_path_ni": "04_construction/man_con/con_file_ni.csv",
         },
         "years": {"survey_year": 2022},
@@ -43,7 +47,7 @@ def config():
         },
         "2022_mappers": {
             "mappers_version": "v1",
-            "postcodes_mapper": "pcodes_2022.csv",
+            "postcode_mapper": "pcodes_2022.csv",
             "itl_mapper_path": "itl_2022.csv",
         },
         "mapping_paths": {
@@ -74,8 +78,12 @@ def test_get_paths(config):
         "ni_full_responses_path": "03_northern_ireland/2021/TEST_ni.csv",
         "manual_outliers_path": "07_outliers/man_out/man_out.csv",
         "manual_imp_trim_path": "06_imputation/man_trim/trim_qa.csv",
-        "all_data_construction_file_path": "04_construction/man_con/construction_file.csv",
-        "postcode_construction_file_path": "04_construction/man_con/postcode_construction_file.csv",
+        "all_data_construction_file_path": (
+            "04_construction/man_con/construction_file.csv"
+        ),
+        "postcode_construction_file_path": (
+            "04_construction/man_con/postcode_construction_file.csv"
+        ),
         "construction_file_path_ni": "04_construction/man_con/con_file_ni.csv",
         "year": 2022,
         "berd_path": "R:/DAP_emulation/2022_surveys/BERD/",
@@ -130,7 +138,7 @@ def test_create_mapping_config(config):
     """Test create_mapping_config function."""
 
     expected_mapping_dict = {
-        "postcodes_mapper": "R:/DAP_emulation/2022_surveys/mappers/v1/pcodes_2022.csv",
+        "postcode_mapper": "R:/DAP_emulation/2022_surveys/mappers/v1/pcodes_2022.csv",
         "itl_mapper_path": "R:/DAP_emulation/2022_surveys/mappers/v1/itl_2022.csv",
         "qa_path": "R:/DAP_emulation/2022_surveys/BERD/05_mapping/mapping_qa",
     }
